+++ src/derive/readlen/compute.rs
@@ -15,7 +15,7 @@ pub struct DerivedReadlenResult {
     pub consensus_read_length: Option<usize>,
 
     /// The majority vote percentage of the consensus read length, if available.
-    pub majority_pct_detected: f64,
+    pub majority_pct_detected: f32,
 
     /// Status of the evidence that supports (or does not support) this
     /// read length, if available.
@@ -27,7 +27,7 @@ impl DerivedReadlenResult {
     pub fn new(
         succeeded: bool,
         consensus_read_length: Option<usize>,
-        majority_pct_detected: f64,
+        majority_pct_detected: f32,
         evidence: Vec<(usize, usize)>,
     ) -> Self {
         DerivedReadlenResult {
@@ -45,7 +45,7 @@ impl DerivedReadlenResult {
 pub fn predict(
     read_lengths: HashMap<usize, usize>,
     num_samples: usize,
-    majority_vote_cutoff: f64,
+    majority_vote_cutoff: f32,
 ) -> Result<DerivedReadlenResult, anyhow::Error> {
     if num_samples == 0 {
         bail!("No read lengths were detected in the file.");
@@ -59,7 +59,7 @@ pub fn predict(
     let max_count = read_lengths[0].1;
 
     let consensus_read_length = max_read_length;
-    let majority_detected = max_count as f64 / num_samples as f64;
+    let majority_detected = max_count as f32 / num_samples as f32;
 
     let mut result =
         DerivedReadlenResult::new(false, None, majority_detected * 100.0, read_lengths);
